conflictコマンドの追加
