+++ README.md
@@ -25,6 +25,7 @@ jpug-doc-tool サブコマンド
 サブコマンドには以下があります。
 
 - `extract` 英文、日本語文の抽出
+- `conflict` 同一英文で異なる日本語訳の抽出
 - `replace` 英文、日本語文の置き換え
 - `mtreplace` 機械翻訳の置き換え
 - `list` 英文、日本語文の出力
